safari-534.53-branch/LayoutTests/java/lc3/ConvertJSObject/ToInt-002-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion:  JavaScript Object to int

FAIL TEST_CLASS.ambiguous( MyFunction ) +'' should be INT. Was BYTE.
FAIL TEST_CLASS.ambiguous( MyOtherFunction ) +'' should be INT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new MyObject(12345) ) +'' should be INT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new MyOtherObject("12345") ) +'' should be INT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new AnotherObject("12345") ) +'' should be INT. Was BYTE.
PASS successfullyParsed is true

TEST COMPLETE
 
